refactor(useInputs): useInputs를 공용화하고 cardNumber 커스텀 훅에서 autoFocus 처리
