+++ src/hooks/useInputs.ts
@@ -1,14 +1,8 @@
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 import { ErrorDetail } from '../types/error';
 import { CardNumberKey } from '../types/card';
 import { INITIAL_ERROR_VALUE } from '../constants/error';
 import { convertArrayIntoObject } from '../utils/util';
-
-const NEXT_INPUT_FIELD = {
-  first: 'second',
-  second: 'third',
-  third: 'fourth',
-} as const;
 
 interface ValidateProps {
   onChange: (value: string) => ErrorDetail;
@@ -19,21 +13,9 @@ const useInputs = <T extends Record<string, string>>(initialValue: T, validate: 
   const initialErrorStatus = convertArrayIntoObject(Object.keys(initialValue), INITIAL_ERROR_VALUE);
   const [value, setValue] = useState(initialValue);
   const [errorInfo, setErrorInfo] = useState(initialErrorStatus);
-  const inputRefs = {
-    first: useRef<HTMLInputElement>(null),
-    second: useRef<HTMLInputElement>(null),
-    third: useRef<HTMLInputElement>(null),
-    fourth: useRef<HTMLInputElement>(null),
-  };
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>, name: string) => {
     const validationResult = validate.onChange(e.target.value);
-    if (e.target.value.length === e.target.maxLength && !validationResult.isError) {
-      const key = name as keyof typeof NEXT_INPUT_FIELD;
-      if (NEXT_INPUT_FIELD[key]) {
-        inputRefs[NEXT_INPUT_FIELD[key]].current?.focus();
-      }
-    }
     setErrorInfo({
       ...errorInfo,
       [name]: {
@@ -66,7 +48,7 @@ const useInputs = <T extends Record<string, string>>(initialValue: T, validate: 
     handleChange,
     handleBlur,
     errorInfo,
-    inputRefs,
+    setErrorInfo,
   };
 };
